+++ js/main.js
@@ -1,10 +1,34 @@
 import {
+  activateMapFilters,
+  deactivateMapFilters
+} from './map-filters.js';
+
+import {
+  activateForm,
+  deactivateForm
+} from './form.js';
+
+import {
 } from './data.js';
 
 import {
-} from './form-validation.js';
-
-import {
+  map
 } from './map.js';
+
+
+const activatePage = () => {
+  activateMapFilters();
+  activateForm();
+};
+
+const deactivatePage = () => {
+  deactivateMapFilters();
+  deactivateForm();
+};
+
+deactivatePage();
+map.on('load', () => {
+  activatePage();
+});
